Move terraform to common json parser
